LayoutTests/imported/w3c/web-platform-tests/domxpath/resolver-callback-interface-cross-realm.tentative-expected.txt



PASS XPathNSResolver is cross-realm plain object without 'lookupNamespaceURI' property
PASS XPathNSResolver is cross-realm plain object with non-callable 'lookupNamespaceURI' property
PASS XPathNSResolver is cross-realm non-callable revoked Proxy
FAIL XPathNSResolver is cross-realm callable revoked Proxy assert_equals: expected function "function TypeError() {
    [native code]
}" but got function "function TypeError() {
    [native code]
}"
FAIL XPathNSResolver is cross-realm plain object with revoked Proxy as 'lookupNamespaceURI' property assert_equals: expected function "function TypeError() {
    [native code]
}" but got function "function TypeError() {
    [native code]
}"
